requirements.txt
Django==6.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt
django-cors-headers==4.9.0
pillow==12.1.0
sqlparse==0.5.5
django-environ==0.12.0
